Added option --llpe to run LLPE for intra-module pruning.
LLPE compiles but it is crashing so it's not usable yet.

One problem is that we modified original LLPE to port it to LLVM 5.0
so we might have introduced some bugs. Another problem is that we
don't know the expectations of --llpe when it's is called. One thing
we noticed is that it's expecting a module with main.

We need to figure out how to use only DIE (dead code elimination) and
DSE (dead store elimination) capabilities to use them during our
intra-module pruning.

Not good idea to mix commits but this commit also adds the option
--ipdse so that we can run optionally our interprocedural dead store
elimination.