Removed trivial devirt that only uses types.
Also, commented out the code that update the call graph.  We are
missing at least one case so we don't bother to update the call graph
for now since the LLVM pass manager will re-build it again. We need to
come back to this at some point and update the callgraph on the fly.